# src/swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    { "url": "http://example.com", "description": "", "variables": {} }
  ],
  "info": {
    "version": "86ab7d87-f57c-468d-85c2-11a2fe030de4",
    "title": "ACB Event - Kha",
    "description": "",
    "termsOfService": "",
    "contact": {},
    "license": { "name": "" }
  },
  "paths": {
...
    "/currenUser": {
      "get": {
        "summary": "currentUser",
        "operationId": "currentUser",
        "parameters": [
          {
            "name": "eventID",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": { "type": "number" }
          },
          {
            "name": "api-token",
            "in": "header",
            "required": true,
            "style": "simple",
            "schema": { "type": "string", "example": "{{access_token_user}}" }
          }
        ],
        "responses": {
          "200": {
            "description": "detail",
            "content": {
              "application/json; charset=utf-8": {
                "schema": { "$ref": "#/components/schemas/200" },
                "example": {
                  "code": 1,
                  "data": {
                    "id": 2,
                    "name": "Nguyễn văn Ab22",
                    "phone": "0947883647",
                    "roles": "[]",
                    "email": "test1@gmail.com"
                  },
                  "message": "ok"
                }
              }
            }
          }
        },
        "tags": []
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "200": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "description": "", "default": 1 },
          "data": {
            "type": "string",
            "description": "",
            "default": "CREATE_SUCCESS"
          },
          "message": { "type": "string", "description": "", "default": "ok" }
        }
      }
    }
  },
  "security": [],
...
}